pyproject.toml
[project]
name = "agent-fly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.8.0",
    "assemblyai==0.40.2",
    "asyncio-throttle>=1.0.2",
    "beautifulsoup4>=4.12.0",
    "chunkr-ai",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "crawl4ai>=0.2.0",
    "docx2markdown==0.1.1",
    "duckduckgo-search>=3.9.0",
    "faiss-cpu>=1.7.4",
    "fastmcp==2.7.0",
    "ffmpeg==7.1.1",
    "google-generativeai>=0.3.0",
    "google-search-results>=2.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.20",
    "langchain-google-genai>=0.0.5",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.40",
    "loguru==0.7.3",
    "nltk>=3.8.0",
    "numpy>=1.24.0",
    "openai==1.75.0",
    "opencv-python==4.11.0.86",
    "pandas==2.2.3",
    "pdfplumber>=0.9.0",
    "playwright>=1.40.0",
    "pypdf2>=3.0.0",
    "python-dotenv==1.1.0",
    "python-pptx==1.0.2",
    "requests>=2.31.0",
    "retry==0.9.2",
    "scenedetect==0.6.6",
    "selenium>=4.15.0",
    "sentence-transformers>=2.2.2",
    "serpapi>=0.1.0",
    "spacy>=3.7.0",
    "sqlite3",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tqdm>=4.65.0",
    "typing-extensions>=4.0.0",
    "xmltodict==0.14.2",
    "yt-dlp",
]
